+++ src/main/java/am/shavigh/api/repo/BiblesBookRepo.java
@@ -3,4 +3,4 @@ package am.shavigh.api.repo;
 import am.shavigh.api.model.bibles.BibleBooks;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BiblesBookRepo extends JpaRepository<BibleBooks, Integer> {}
+public interface BiblesBookRepo extends JpaRepository<BibleBooks, Long> {}
